Move affix presets to affix section
